requirements.txt
fastapi
uvicorn
pydantic
asyncpg
alembic
sqlalchemy
psycopg2-binary
python-jose
redis
pydantic[email]
passlib
python-multipart
slowapi
starlette